requirements.txt
Django==2.2.1
djangorestframework==3.9.3
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0

appnope==0.1.0
autopep8==1.4.4
backcall==0.1.0
decorator==4.4.0
ipdb==0.12
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7